requirements.txt
fsspec
pytz
tzdata
pefile
pyinstaller-hooks-contrib
certifi
setuptools
pip
attrs
packaging
pyarrow
Pillow
pyinstaller
cachetools
fonttools
protobuf
lxml
rsa
typing-extensions
zeep
pyodbc
matplotlib
prettytable
platformdirs
idna
pyspark
openpyxl
charset-normalizer
pyparsing
requests
google-auth
google-api-core
google-cloud-storage
pycodestyle
psycopg2
python-dateutil
google-resumable-media
google-cloud-core
mysql-connector
mysqlclient
SQLAlchemy
autopep8
greenlet
pandas
googleapis-common-protos
botocore
boto3
urllib3
numpy
six
requests-file
google-crc32c
kiwisolver
schedule
et-xmlfile
contourpy
jmespath
python-dotenv
requests-toolbelt
altgraph
seaborn
cycler
py4j
tabulate
isodate
s3transfer
pyasn1
pyasn1-modules
wcwidth
pywin32-ctypes
mysql
